# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ESPN Fantasy Football Trade Simulator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: espn
    description: ESPN Fantasy Football operations
  - name: trades
    description: Trade simulation operations
  - name: auth
    description: User account registration, login, and settings
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /espn/connect:
    post:
      operationId: connectEspn
      tags: [espn]
      summary: Connect ESPN account
      description: Store ESPN credentials (s2 cookie and SWID) to access private leagues
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EspnConnectBody"
      responses:
        "200":
          description: Connection verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EspnConnectResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /espn/leagues:
    get:
      operationId: getLeagues
      tags: [espn]
      summary: Get user leagues
      description: Returns all fantasy football leagues for the connected account
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of leagues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/League"
        "401":
          description: Not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /espn/leagues/{leagueId}/teams:
    get:
      operationId: getLeagueTeams
      tags: [espn]
      summary: Get teams in a league
      description: Returns all teams in the specified league with their rosters
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: string
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
        - name: season
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of teams with rosters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /trades/simulate:
    post:
      operationId: simulateTrade
      tags: [trades]
      summary: Simulate a multi-team trade
      description: >
        Simulate a trade involving 2 or more teams using an explicit origin-to-destination
        matrix. Each transfer specifies exactly which player moves from which team to which team,
        allowing any topology (A→B, A→C, B→A, etc.) rather than a fixed circular chain.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulateTradeBody"
      responses:
        "200":
          description: Trade simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeSimulationResult"
        "400":
          description: Invalid trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /trades/saved:
    get:
      operationId: getSavedTrades
      tags: [trades]
      summary: Get saved trade simulations
      description: Returns all saved trade scenarios for the given session
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of saved trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedTrade"
    post:
      operationId: saveTrade
      tags: [trades]
      summary: Save a trade simulation
      description: Save a trade simulation result for later reference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveTradeBody"
      responses:
        "201":
          description: Trade saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedTrade"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /trades/saved/{tradeId}/refresh:
    post:
      operationId: refreshSavedTrade
      tags: [trades]
      summary: Refresh trade scores using current ESPN player values
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: integer
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated saved trade with fresh scores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedTrade"
        "401":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Trade not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: ESPN fetch failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /trades/saved/{tradeId}:
    delete:
      operationId: deleteSavedTrade
      tags: [trades]
      summary: Delete a saved trade
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "200":
          description: Account created and token returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Sign in to an existing account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Authenticated token returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the current logged-in user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/settings:
    patch:
      operationId: updateUserSettings
      tags: [auth]
      summary: Update user preferences (e.g. league warning visibility)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    EspnConnectBody:
      type: object
      properties:
        espnS2:
          type: string
          description: ESPN S2 cookie value
        swid:
          type: string
          description: ESPN SWID cookie value
      required:
        - espnS2
        - swid

    EspnConnectResponse:
      type: object
      properties:
        sessionId:
          type: string
        username:
          type: string
        connected:
          type: boolean
      required:
        - sessionId
        - connected

    League:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        season:
          type: integer
        currentWeek:
          type: integer
        teamCount:
          type: integer
        scoringType:
          type: string
      required:
        - id
        - name
        - season
        - currentWeek
        - teamCount

    Player:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        position:
          type: string
        nflTeam:
          type: string
        points:
          type: number
        projectedPoints:
          type: number
        tradeValue:
          type: number
        isStarter:
          type: boolean
        injuryStatus:
          type: ["string", "null"]
      required:
        - id
        - name
        - position
        - nflTeam
        - points
        - projectedPoints
        - tradeValue
        - isStarter

    Team:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        abbreviation:
          type: string
        ownerName:
          type: string
        wins:
          type: integer
        losses:
          type: integer
        ties:
          type: integer
        pointsFor:
          type: number
        pointsAgainst:
          type: number
        totalTradeValue:
          type: number
        roster:
          type: array
          items:
            $ref: "#/components/schemas/Player"
      required:
        - id
        - name
        - ownerName
        - wins
        - losses
        - ties
        - pointsFor
        - pointsAgainst
        - totalTradeValue
        - roster

    PlayerTransfer:
      type: object
      description: >
        A single player movement from one team to another. Using an explicit
        origin/destination pair instead of a circular chain lets any team send
        any player to any other participating team in the same trade.
      properties:
        playerId:
          type: string
          description: The ID of the player being transferred
        fromTeamId:
          type: string
          description: The team giving this player up
        toTeamId:
          type: string
          description: The team receiving this player
      required:
        - playerId
        - fromTeamId
        - toTeamId

    SimulateTradeBody:
      type: object
      properties:
        sessionId:
          type: string
        leagueId:
          type: string
        transfers:
          type: array
          description: >
            Explicit origin-to-destination matrix. Each entry is one player movement.
            Participating teams are derived from the union of all fromTeamId/toTeamId values.
          items:
            $ref: "#/components/schemas/PlayerTransfer"
          minItems: 1
        teams:
          type: array
          items:
            $ref: "#/components/schemas/Team"
      required:
        - sessionId
        - leagueId
        - transfers
        - teams

    RosterOverflow:
      type: object
      description: >
        Present on a TeamTradeResult when the team's post-trade roster would exceed
        their pre-trade size. The user must select `excess` players to drop before
        the trade can be finalized.
      properties:
        excess:
          type: integer
          description: Number of players this team must drop to reach a legal roster size
      required:
        - excess

    TeamTradeResult:
      type: object
      properties:
        teamId:
          type: string
        teamName:
          type: string
        ownerName:
          type: string
        playersGiven:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        playersReceived:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        rosterBefore:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        rosterAfter:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        tradeValueBefore:
          type: number
        tradeValueAfter:
          type: number
        tradeValueChange:
          type: number
        verdict:
          type: string
          description: "win, loss, or neutral"
        grade:
          type: string
          description: "Letter grade A+ through F for this team's side of the trade"
        score:
          type: number
          description: "Numeric score 0-100 for this team's side of the trade"
        gradeRationale:
          type: string
          description: "Short explanation of the grade"
        rosterOverflow:
          $ref: "#/components/schemas/RosterOverflow"
          description: >
            Only present when this team would exceed their pre-trade roster size.
            Absent (undefined) when the team's roster size is legal post-trade.
      required:
        - teamId
        - teamName
        - ownerName
        - playersGiven
        - playersReceived
        - rosterBefore
        - rosterAfter
        - tradeValueBefore
        - tradeValueAfter
        - tradeValueChange
        - verdict
        - grade
        - score
        - gradeRationale

    TradeSimulationResult:
      type: object
      properties:
        leagueId:
          type: string
        teamResults:
          type: array
          items:
            $ref: "#/components/schemas/TeamTradeResult"
        overallBalance:
          type: number
          description: Sum of all value changes (should be ~0 for fair trades)
        summary:
          type: string
        hasRosterOverflow:
          type: boolean
          description: >
            True when at least one participating team's post-trade roster exceeds
            their pre-trade size. The frontend must resolve all overflows (drops)
            before the trade can be saved.
        leagueWarnings:
          type: array
          description: >
            Human-readable descriptions of league rule violations produced by
            this trade (e.g. roster overflow, positional limits exceeded).
            Show a soft warning banner to the user when this array is non-empty.
          items:
            type: string
      required:
        - leagueId
        - teamResults
        - overallBalance
        - summary
        - hasRosterOverflow
        - leagueWarnings

    UserProfile:
      type: object
      description: Logged-in user account details
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        showLeagueWarnings:
          type: boolean
          description: Whether to show the Soft Warning Sandbox banner on the results screen
        createdAt:
          type: string
          description: ISO timestamp
      required:
        - id
        - email
        - showLeagueWarnings
        - createdAt

    AuthToken:
      type: object
      properties:
        token:
          type: string
          description: JWT bearer token — store locally and include in Authorization header
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user

    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          description: Minimum 8 characters
      required:
        - email
        - password

    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    UpdateSettingsBody:
      type: object
      properties:
        showLeagueWarnings:
          type: boolean
          description: Set to false to hide the Soft Warning Sandbox banner permanently
      required:
        - showLeagueWarnings

    SaveTradeBody:
      type: object
      properties:
        sessionId:
          type: string
        leagueId:
          type: string
        name:
          type: string
        result:
          $ref: "#/components/schemas/TradeSimulationResult"
        transfers:
          type: array
          description: The transfer matrix used to generate this result, stored for future refresh
          items:
            $ref: "#/components/schemas/PlayerTransfer"
      required:
        - sessionId
        - leagueId
        - name
        - result
        - transfers

    SavedTrade:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        leagueId:
          type: string
        name:
          type: string
        result:
          $ref: "#/components/schemas/TradeSimulationResult"
        transfers:
          type: array
          description: The original transfer matrix, used to re-simulate on refresh
          items:
            $ref: "#/components/schemas/PlayerTransfer"
        lastRefreshedAt:
          type: string
          description: ISO timestamp of when scores were last recalculated from live ESPN data
        createdAt:
          type: string
      required:
        - id
        - sessionId
        - leagueId
        - name
        - result
        - transfers
        - lastRefreshedAt
        - createdAt